backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anyio>=4.13.0",
    "arq>=0.28.0",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "cloudinary>=1.44.2",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "firebase-admin>=7.4.0",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "pydantic[email]>=2.13.3",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.27",
    "redis>=5.3.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "httpx>=0.28.1",
    "isort>=8.0.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
]
